Move some functions to tx_verify
