+++ src/main.cpp
@@ -532,24 +532,6 @@ bool IsStandardTx(const CTransaction& tx)
     return true;
 }
 
-bool IsFinalTx(const CTransaction &tx, int nBlockHeight, int64_t nBlockTime)
-{
-    AssertLockHeld(cs_main);
-    // Time based nLockTime implemented in 0.1.6
-    if (tx.nLockTime == 0)
-        return true;
-    if (nBlockHeight == 0)
-        nBlockHeight = nBestHeight;
-    if (nBlockTime == 0)
-        nBlockTime = GetAdjustedTime();
-    if ((int64_t)tx.nLockTime < ((int64_t)tx.nLockTime < LOCKTIME_THRESHOLD ? (int64_t)nBlockHeight : nBlockTime))
-        return true;
-    for (auto const& txin : tx.vin)
-        if (!txin.IsFinal())
-            return false;
-    return true;
-}
-
 
 int CMerkleTx::SetMerkleBranch(const CBlock* pblock)
 {
